tidying up code currently it cannot compile added files and folders for modifiers and entities eg players the modifiers are all function pointers used to manipluate the map
