--- src/map_parser/fake_wall_texture.txt
@@ -1,1 +1,0 @@
-qwe
