+++ src/map_parser/map_tools.c
@@ -1,0 +1,101 @@
+#include "cubed.h"
+
+int	map_skip_over_modifiers(char *content)
+{
+	int	i;
+//	char	temp[3];
+	
+	i = -1;
+//	strncpy(temp, content, 2);
+//	temp[2] = '\0';
+	while (++i < sizeof(g_mapsymbols) / 8)
+	{
+		if (ft_strncmp(g_mapsymbols[i], content, mod_strlen(g_mapsymbols[i])) == 0)
+			return (1);
+//		if (strcmp(g_mapsymbols[i], temp) == 0)
+//			return (1);
+//		printf("%s\n", g_mapsymbols[i]);
+	}
+	return (0);
+}
+
+int	map_starting_tile(char *content)
+{
+	int	i;
+	
+//	if (!content)
+	i = 0;
+	while (ft_isspace(content[i]))
+		i++;
+	if(content[i] == '1')
+		return (1);
+	return (0);
+}
+
+
+void	replace_tabs(t_list *curr)
+{
+	int	i;
+	int	j;
+	char	*line;
+
+	while (curr != NULL && curr->content != NULL)
+	{
+		line = (char *)curr->content;
+		i = 0;
+		while (line[i] != '\0')
+		{
+			if (line[i] == '\t')
+			{
+				ft_memmove(&line[i + TAB_SIZE], &line[i + 1], ft_strlen(&line[i + 1]) + 1);
+				j = 0;
+				while (j < TAB_SIZE)
+					line[i + j++] = ' ';
+				i += TAB_SIZE;
+			}
+			else
+				i++;
+		}
+		curr = curr->next;
+	}
+}
+
+void	map_print(t_map *map)
+{
+	int	i;
+
+	i = 0;
+	while (i < (map->width * map->height))
+	{
+		if (i % map->width == 0)
+			printf("\n");
+		printf("%d", map->tiles[i].type);
+		i++;
+	}
+}
+
+
+void	remove_empty_lines(t_list **raw_map_file)
+{
+	t_list	*curr;
+	t_list	*temp;
+
+//	int lines = (intptr_t)raw_map_file[0]->content;
+	temp = *raw_map_file;
+	*raw_map_file = temp->next;
+	free(temp);
+	curr = *raw_map_file;
+	while(curr != NULL && curr->next != NULL) 
+	{
+		if (map_starting_tile((char *)curr->content) == 1)
+			break ;
+		if (is_empty_line(curr->next->content) == 1)
+		{
+			temp = curr->next;
+			curr->next = temp->next;
+			ft_lstdelone(temp, free);
+		}
+		else
+			curr = curr->next;
+	}
+}
